CLEANUP: moved in proper folders and namespacing
