+++ generators/app/templates/src/StarterKit/Api/Models/Examples/Example.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace StarterKit.Api.Models
+namespace StarterKit.Api.Models.Examples
 {
     public class Example : ModelBase
     {
